binary: move to strace/internal
pkg/binary leaked from gVisor. It is an optimization of encoding/binary.
I don't think we necessarily care about that optimization in u-root, but
let's keep it around for now.
