+++ pkg/strace/internal/abi/abi_unix.go
@@ -18,7 +18,7 @@ import (
 	"fmt"
 	"syscall"
 
-	"github.com/u-root/u-root/pkg/binary"
+	"github.com/u-root/u-root/pkg/strace/internal/binary"
 	"golang.org/x/sys/unix"
 )
