+++ pkg/strace/internal/binary/binary.go
