+++ pkg/strace/socket.go
@@ -18,8 +18,8 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/u-root/u-root/pkg/binary"
 	"github.com/u-root/u-root/pkg/strace/internal/abi"
+	"github.com/u-root/u-root/pkg/strace/internal/binary"
 	"golang.org/x/sys/unix"
 )
